+++ Changes.md
@@ -4,6 +4,7 @@ Version 3.10.1.3 - TBR
 -----------------------------
   1. Add 'aria-controls' (https://www.w3.org/TR/wai-aria-practices/examples/disclosure/disclosure-faq.html) - CONTRIB-8637.
   2. Add caching to activity meta information in the hope of improving performance.
+  3. Add 'aria-live=polite' to the section content and 'aria-controls' to the open / close all buttons - CONTRIB-8637.
 
 Version 3.10.1.2 - 02/07/2021
 -----------------------------